Modify the general architecture of permissions system and where they are checked

+++ src/domain/card/CardRepository.ts
@@ -5,12 +5,8 @@ import {CardDao} from './CardDao.js'
 import {AuthorIdentification} from './AuthorIdentification.js'
 import {PersistenceFactory} from '../../implementations/persistence/PersistenceFactory.js'
 import {Authorization} from '../authorization/Authorization.js'
-import {RequesterIdentification} from '../authorization/permission/RequesterIdentification.js'
-import {CreateCard} from '../authorization/permission/permissions/CreateCard.js'
 import {DeleteCard} from '../authorization/permission/permissions/DeleteCard.js'
-import {UpdateCard} from '../authorization/permission/permissions/UpdateCard.js'
-import {TransferCard} from '../authorization/permission/permissions/TransferCard.js'
-import {User} from '../user/User.js'
+import {UserPermissions} from '../authorization/UserPermissions.js'
 
 export class CardRepository {
 
@@ -20,13 +16,12 @@ export class CardRepository {
         this.dao = PersistenceFactory.getCardDao()
     }
 
-    async add(card: Card, requesterId: RequesterIdentification) {
-        await Authorization.assert(requesterId).can(CreateCard, card)
+    async add(card: Card) {
         await this.dao.insert(card)
     }
 
-    async delete(card: Card, requesterId: RequesterIdentification) {
-        await Authorization.assert(requesterId).can(DeleteCard, card)
+    async delete(card: Card, userPermissions: UserPermissions) {
+        Authorization.assertUserWithPermissions(userPermissions).can(DeleteCard, card)
         await this.dao.delete(card.getId())
     }
 
@@ -34,15 +29,8 @@ export class CardRepository {
         return this.dao.findById(id)
     }
 
-    async updateData(card: Card, requesterId: RequesterIdentification) {
-        await Authorization.assert(requesterId).can(UpdateCard, card)
+    async update(card: Card) {
         return this.dao.update(card)
-    }
-
-    async transfer(card: Card, user: User, requesterId: RequesterIdentification) {
-        const transferredCard = card.transferTo(user)
-        await Authorization.assert(requesterId).can(TransferCard, card)
-        return this.dao.update(transferredCard)
     }
 
     async findByLabelling(labelling: Labelling): Promise<Card[]> {
